Eliminate history-related import cycles
